Pass quick_search_config.yml file through ERB processor
The quick_search_config.yml file contains some properties that
might differ on a dev server vs. a production server. For example,
the Best Bets "solr_url" property might point to a dev Solr instance
from a dev server, and a production Solr instance on a production
server.

Is it often convenient to provide such information via environment
variables that are passed into Rails. This pull request modifies
the processing of the config/quick_search_config.yml file to pass
it through the ERB template processor. ERB directives such as
"<%= ENV['SOLR_URL'] %>" can then be added to the file, to enable
environment variables to be utilized.