LIBSEARCH-15. For failed searches, allow use of "loaded_link" method
There are currently two mechanisms for specifying the "loaded_link"
(a direct link to the site being searched) in the GUI:

1) Via the "loaded_link" method on the searcher

2) Via a "<searcher_name>_search.loaded_link" entry in the I18N locale
files (i.e., config/locales/en.yml).

Having a "loaded_link" property in the I18N locale files, and a
"loaded_link" method in the searcher is confusing, as it is unclear
without examining the code, which instance is used where. It is even
more problematic, because the return value from the "loaded_link" method
may contain additional processing that is not possible with the
I18N "loaded_link" property.

In the existing implementation, the result of the "loaded_link" method
is displayed in the GUI when there is a successful search with results.
The "loaded_link" property from the I18N locale files is displayed when
a searcher error occurs.

From a larger perspective, having the "loaded_link" property in the
I18N locale files is itself an issue, because it makes it extremely
difficult to configure the property on a per-server basis. For example,
we have a "search-dev" machine whose searchers point to other
development machines, while our "search" production machine has
searchers pointing to other production machines. Since the I18N locale
files can't be configured via server environment variables, the file
needs to be manually changed on either the production or dev machine
(depending on which version of the I18N locale files are checked into
the Git repository). When the "loaded_link" method is used, the method
can use the searcher configuration file in "config/searchers" directory
to determine the URL, and that configuration file can be parameterized
to use environment variables.

In the existing implementation the "search_all_in_threads" method of
SearcherConcern class returns an instance variable (named after the
searcher) that contains either the searcher object itself (if no error
occurs), or a StandardError (if an error occurs). The presence of
a StandardError in the instance variable triggers the appropriate
display in the view templates, but does not include a reference to the
searcher itself.

This change makes it possible for the "loaded_link" method of the
searcher to be called even when there is an error. It does this by
wrapping the StandardError returned by the "search_all_in_threads"
method in a "SearcherError" class (which itself extends StandardError),
providing a accessor for retrieving the searcher. If an error has
occurred, the searcher is available to the ERB template, so that
the "loaded_link" method can be called.

In order to preserve the existing behavior of utilizing the
"loaded_link" property from the I18N locale file, logic has been placed
in the ERB templates to prefer the I18N property (if it exists) over
the "loaded_link" method of the searcher for failed searches.

https://issues.umd.edu/browse/LIBSEARCH-15