LIBSEARCH-15. Added "no_results_link" method for searchers
In the existing implementation, the "no_results_link" can only be
specified in the I18N locale files.

Having the "no_results_link" property in the I18N locale files makes it
difficult to configure the value on a per-server basis. For example,
we have a "search-dev" machine whose searchers point to other
development machines, while our "search" production machine has
searchers pointing to other production machines. Since the I18N locale
files can't be configured via server environment variables, the file
needs to be manually changed on either the production or dev machine
(depending on which version of the I18N locale files are checked into
the Git repository).

In this implementation, a "no_results_link" method has been added to
the QuickSearch::Searcher base class. The class provides a default
implementation that preserves the existing behavior of retrieving
the "no_results_link" value from the I18N locale files, but also
adds the ability to retrieve the value from the searcher configuration
file.

Allowing the "no_results_link" value to be specified in the searcher
configuration file allows greater flexibility in specifying the link,
as the searcher configuration file can be parameterized using
environment variables.

https://issues.umd.edu/browse/LIBSEARCH-15