Moved "module_link" handling into Searcher base class
Removed duplication of "module_link" handling in the "_module.html.erb"
and "_module_with_paging.html.erb" view helpers by moving the logic
into a "module_link_on_error" class method on QuickSearch::Searcher.

The "module_link_on_error" method is a class method because when an
error occurs, we might not have access to the searcher instance
that failed.

The "module_link_on_error" method preserves the existing behavior of
preferring the I18N property (if it exists) over the "loaded_link"
method of the searcher.