LIBSEARCH-112. Fixed "no_results_link" handling for I18N fallback
In the development environment, the I18n functionality does not
"fallback" to a default if the given I18N key is not found.

In the production environment, the Rails default is to have
"config.i18n.fallback" set to true, in which case a missing
I18N key causes "no translation" text to be returned. This was causing
the value from the searcher configuration file to be ignored, even
if the I18N keys did not exist.

This change modifies the "no_results_link" method to verify that the
I18N key (and I18N default key, if given) actually exist, before using
them as the return value. This ensures that the value from the
searcher configuration file will be used if I18N keys do not exist.

Added additional tests to verify the expected behavior.

https://issues.umd.edu/browse/LIBSEARCH-112