LIBSEARCH-128. Removed "primary_searcher_config" variable
Removed "primary_searcher_config" variable as it is causing all
single searcher HTTP searches to immediately fail, due to an attempt
to treat a String as a Hash.

Replaced usages of the "primary_searcher_config" variable with the
"searcher_config" variable.

This prevent a downstream error when using searchers that use a
"loaded_link" method.

https://issues.umd.edu/browse/LIBSEARCH-128