Replaced "primary_searcher_config" variable with "searcher_config"
The "primary_searcher_config" variable is causing all single searcher
searches to immediately fail. The variable is initialized as a String,
and never assigned any other value. When a single searcher searcher is
performed, the "has_key?" method is called, which will always fail,
because the variable is a String, not a Hash.

It is unclear what the original intent of the "primary_searcher_config"
variable was, but it seems similar to the existing "searcher_config"
variable.

Replacing the "primary_searcher_config" variable with "searcher_config"
enables single searcher searches to succeed.

This issue was uncovered while working with searchers that use a
"loaded_link" method to return the URL to the native search interface,
instead of the "loaded_link" property in the I18n en.yml file
(see Pull Request #23). When using such a searcher, the above issue
causes the rendering of the search results page to fail, because the
searcher in the SearcherError object that gets created is "nil".