Added search_cfg nil check for single searcher search page
Fixed an issue where the single searcher search page displays an error
for searchers without an explicit "config/<searcher>_config.yml"
configuration file.

The cause of the error is not handling the possibility that the
"search_cfg" variable may be nil. This is likely just an oversight,
as a nil check is performed in other places in the method.

Modified the code to provide "@common_searches" with a default value,
which can then be overwritten if the "search_cfg" variable contains
the "common_searches" key.