LIBSEARCH-155. Pass Best Bet keywords through "filter_query" method
When loading the Best Bet keywords, pass them through the "filter_query"
method in the "QuickSearch::QueryFilter" concern, so that they will be
matched with a query term passed  through the same method. This takes
care of things like a dash (hyphen) being removed from the query term.

The implementation was somewhat complicated by the fact that the
"filter_query" method is a private method on a Rails Concern. This was
handled by creating a helper class in the initializer that delegates to
the concern. It is not clear if this is the best way to accomplish this,
but it does appear to work.

https://issues.umd.edu/browse/LIBSEARCH-155